🐛 HOTFIX: Resolve failing filter category test
The test (filter.spec.js) failed at times especially in CI,
because it wasn't accounting for the svelte fade out transition duration.
Meaning: While the opacity of the soon-to-be-removed flags faded,
they sometimes still got counted as part of the flag array
and therefore the assignment failed (at times)
The randomness of failure was *likely* accounted by the execution time the transition needs to complete.

Removing the transition will barly be noticable and fix the test,
which I'll account as the "greater good".